% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripserq-package.R
\docType{package}
\name{ripserq-package}
\alias{ripserq}
\alias{ripserq-package}
\title{ripserq: Concise Summary of What the Package Does}
\description{
More about what it does (maybe more than one line).
}
\author{
\strong{Maintainer}: Your Name \email{your@email.com}

}
\keyword{internal}
